Open in Finder (#101)
* debug

* rahuls changes + use app sandbox for mac

* fix compile error w appkit on ios

+++ CanvasPlusPlayground/Features/Files/CourseFileService.swift
@@ -6,12 +6,36 @@
 //
 
 import Foundation
+#if os(macOS)
+import AppKit
+#endif
 
 struct CourseFileService {
     
     private static let fileManager: FileManager = .default
-    private static var documentsURL: URL {
-        fileManager.urls(for: .documentDirectory, in: .userDomainMask).first!
+    private static var rootURL: URL? {
+        if let bundleId = Bundle.main.bundleIdentifier {
+            let root = URL.applicationSupportDirectory.appendingPathComponent(bundleId)
+            
+            if fileManager.fileExists(atPath: root.path) {
+                return root
+            } else {
+                do {
+                    try fileManager.createDirectory(at: root, withIntermediateDirectories: true, attributes: nil)
+                    return root
+                } catch {
+                    print("Failure creating directory to root.")
+                    return nil
+                }
+            }
+            
+        } else {
+            print("Failure getting bundle identifier")
+            return nil
+        }
+    }
+    private static var coursesURL: URL? {
+        rootURL?
             .appendingPathComponent(StorageKeys.accessTokenValue)
             .appendingPathComponent("courses")
     }
@@ -29,7 +53,7 @@ struct CourseFileService {
             throw FileError.unsupportedFileType
         }
         
-        let fileURL = self.pathWithFolders(foldersPath: folderIds, courseId: courseId, fileId: file.id, type: type)
+        let fileURL = try self.pathWithFolders(foldersPath: folderIds, courseId: courseId, fileId: file.id, type: type)
         let parentDirURL = fileURL.deletingLastPathComponent()
                 
         try Self.fileManager.createDirectory(
@@ -55,10 +79,10 @@ struct CourseFileService {
         foldersPath: [String],
         localCopyReceived: (Data?) -> Void
     ) async throws -> Data {
-        let fileLoc = pathWithFolders(foldersPath: foldersPath, courseId: course.id, fileId: file.id, type: FileType.fromFile(file))
+        let fileLoc = try? pathWithFolders(foldersPath: foldersPath, courseId: course.id, fileId: file.id, type: FileType.fromFile(file))
         
         // Provide local copy meanwhile
-        if Self.fileManager.fileExists(atPath: fileLoc.path()), let data = try? Data(contentsOf: fileLoc) {
+        if let fileLoc, Self.fileManager.fileExists(atPath: fileLoc.path()), let data = try? Data(contentsOf: fileLoc) {
             localCopyReceived(data)
         }
         
@@ -71,11 +95,16 @@ struct CourseFileService {
             
             let content = try Data(contentsOf: tempFileLoc)
             
-            if let file, let url = try? self.saveCourseFile(courseId: course.id, folderIds: foldersPath, file: file, content: content) {
+            do {
+                guard let file else {
+                    throw FileError.fileWasNil
+                }
+                let url = try self.saveCourseFile(courseId: course.id, folderIds: foldersPath, file: file, content: content)
                 print("File successfully saved at \(url.path())")
-            } else {
-                print("Failed to save file.")
+            } catch {
+                print("Failed to save file. \(error)")
             }
+            
             return content
             
         } else {
@@ -83,22 +112,33 @@ struct CourseFileService {
         }
     }
     
+    // MARK: Global
+    
     static func clearAllFiles() throws {
-        let documentURLs = documentsURL
+        guard let fileURL = coursesURL else {
+            throw FileError.directoryInaccessible
+        }
 
         do {
-            let fileURLs = try fileManager.contentsOfDirectory(at: documentURLs, includingPropertiesForKeys: nil)
+            let fileURLs = try fileManager.contentsOfDirectory(at: fileURL, includingPropertiesForKeys: nil)
             
             for fileURL in fileURLs {
                 try fileManager.removeItem(at: fileURL)
                 print("Deleted: \(fileURL.lastPathComponent)")
             }
             
-            print("All files in \(documentURLs.path) have been deleted.")
+            print("All files in \(fileURL.path) have been deleted.")
             
         } catch {
             print("Error deleting files: \(error.localizedDescription)")
         }
+    }
+    
+    /// Opens finder tab at specific directory URL
+    static func showInFinder(fileURL: URL = rootURL ?? .currentDirectory()) {
+        #if os(macOS)
+        NSWorkspace.shared.open(fileURL)
+        #endif
     }
     
     // MARK: Helpers
@@ -116,8 +156,12 @@ struct CourseFileService {
         return tempUrl
     }
     
-    private func pathWithFolders(foldersPath: [String], courseId: String, fileId: String, type: FileType?) -> URL {
-        var pathURL = Self.documentsURL
+    private func pathWithFolders(foldersPath: [String], courseId: String, fileId: String, type: FileType?) throws -> URL {
+        guard let coursesURL = Self.coursesURL else {
+            throw FileError.directoryInaccessible
+        }
+        
+        var pathURL = coursesURL
             .appendingPathComponent(courseId)
             .appendingPathComponent("files")
         for folderId in foldersPath {
@@ -131,7 +175,7 @@ struct CourseFileService {
 
 
 enum FileError: Error {
-    case unsupportedFileType, fileWriteFailed
+    case unsupportedFileType, fileWriteFailed, fileWasNil, directoryInaccessible
     
     var message: String {
         switch self {
@@ -139,6 +183,14 @@ enum FileError: Error {
             return "Unsupported file type"
         case .fileWriteFailed:
             return "File save failed"
+        case .fileWasNil:
+            return "File was nil"
+        case .directoryInaccessible:
+            return "Directory inaccessible"
         }
+    }
+    
+    var description: String {
+        message
     }
 }
